Refactor transaction input data handling by introducing new modules for various blockchain types. Remove deprecated files related to OneInch and Thorchain swap transactions. Update EVM transaction input data to support new structures and improve code organization.
