+++ core/mpc/keysign/txInputData/TxInputDataResolver.ts
@@ -6,13 +6,13 @@ import {
 import { KeysignPayload } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
 import { WalletCore } from '@trustwallet/wallet-core'
 
-export type GetPreSignedInputDataInput<T extends KeysignChainSpecificKey> = {
+export type GetTxInputDataInput<T extends KeysignChainSpecificKey> = {
   keysignPayload: KeysignPayload
   walletCore: WalletCore
   chain: ChainsBySpecific<T>
   chainSpecific: KeysignChainSpecificValueByKey<T>
 }
 
-export type PreSignedInputDataResolver<T extends KeysignChainSpecificKey> = (
-  input: GetPreSignedInputDataInput<T>
-) => Uint8Array<ArrayBufferLike>
+export type TxInputDataResolver<T extends KeysignChainSpecificKey> = (
+  input: GetTxInputDataInput<T>
+) => Promise<Uint8Array<ArrayBufferLike>[]>
