+++ core/mpc/keysign/txInputData/maya.ts
@@ -1,66 +1,49 @@
-import { cosmosFeeCoinDenom } from '@core/chain/chains/cosmos/cosmosFeeCoinDenom'
 import { cosmosGasLimitRecord } from '@core/chain/chains/cosmos/cosmosGasLimitRecord'
 import { getCoinType } from '@core/chain/coin/coinType'
-import { isFeeCoin } from '@core/chain/coin/utils/isFeeCoin'
 import { assertField } from '@lib/utils/record/assertField'
 import { TW } from '@trustwallet/wallet-core'
 import Long from 'long'
 
-import { fromCommCoin } from '../../../types/utils/commCoin'
-import { PreSignedInputDataResolver } from '../PreSignedInputDataResolver'
+import { TxInputDataResolver } from './TxInputDataResolver'
 
-export const getThorPreSignedInputData: PreSignedInputDataResolver<
-  'thorchainSpecific'
-> = ({ keysignPayload, walletCore, chain, chainSpecific }) => {
+export const getMayaTxInputData: TxInputDataResolver<'mayaSpecific'> = async ({
+  keysignPayload,
+  walletCore,
+  chain,
+  chainSpecific,
+}) => {
   const coinType = getCoinType({
     walletCore,
     chain,
   })
 
-  const commCoin = assertField(keysignPayload, 'coin')
+  const coin = assertField(keysignPayload, 'coin')
 
-  const fromAddr = walletCore.AnyAddress.createWithString(
-    commCoin.address,
-    coinType
+  const fromAddr = walletCore.AnyAddress.createBech32(
+    coin.address,
+    coinType,
+    'maya'
   )
 
-  const pubKeyData = Buffer.from(commCoin.hexPublicKey, 'hex')
+  const pubKeyData = Buffer.from(coin.hexPublicKey, 'hex')
 
   let thorchainCoin = TW.Cosmos.Proto.THORChainCoin.create({})
   let message: TW.Cosmos.Proto.Message[]
 
-  if (keysignPayload.memo?.startsWith('merge:')) {
-    const fullDenom = keysignPayload.memo!.toLowerCase().replace('merge:', '')
-
-    message = [
-      TW.Cosmos.Proto.Message.create({
-        wasmExecuteContractGeneric:
-          TW.Cosmos.Proto.Message.WasmExecuteContractGeneric.create({
-            senderAddress: commCoin.address,
-            contractAddress: keysignPayload.toAddress,
-            executeMsg: '{ "deposit": {} }',
-            coins: [
-              TW.Cosmos.Proto.Amount.create({
-                denom: fullDenom,
-                amount: keysignPayload.toAmount,
-              }),
-            ],
-          }),
-      }),
-    ]
-  } else if (chainSpecific.isDeposit) {
+  if (chainSpecific.isDeposit) {
     thorchainCoin = TW.Cosmos.Proto.THORChainCoin.create({
       asset: TW.Cosmos.Proto.THORChainAsset.create({
-        chain: 'THOR',
-        symbol: commCoin.ticker,
-        ticker: commCoin.ticker,
+        chain: 'MAYA',
+        symbol: 'CACAO',
+        ticker: 'CACAO',
         synth: false,
       }),
     })
+
     const toAmount = Number(keysignPayload.toAmount || '0')
     if (toAmount > 0) {
       thorchainCoin.amount = keysignPayload.toAmount
-      thorchainCoin.decimals = new Long(commCoin.decimals)
+      thorchainCoin.decimals = new Long(coin.decimals)
     }
 
     message = [
@@ -74,14 +57,15 @@ export const getThorPreSignedInputData: PreSignedInputDataResolver<
       }),
     ]
   } else {
-    const toAddress = walletCore.AnyAddress.createWithString(
+    const toAddress = walletCore.AnyAddress.createBech32(
       keysignPayload.toAddress,
-      coinType
+      coinType,
+      'maya'
     )
-    if (!toAddress) {
-      throw new Error('invalid to address')
+
+    if (toAddress.description() !== keysignPayload.toAddress) {
+      throw new Error('To address is different from the bech32 address')
     }
-    const coin = fromCommCoin(commCoin)
 
     message = [
       TW.Cosmos.Proto.Message.create({
@@ -89,7 +73,7 @@ export const getThorPreSignedInputData: PreSignedInputDataResolver<
           fromAddress: fromAddr.data(),
           amounts: [
             TW.Cosmos.Proto.Amount.create({
-              denom: isFeeCoin(coin) ? cosmosFeeCoinDenom[chain] : coin.id,
+              denom: coin.ticker.toLowerCase(),
               amount: keysignPayload.toAmount,
             }),
           ],
@@ -102,7 +86,7 @@ export const getThorPreSignedInputData: PreSignedInputDataResolver<
   const input = TW.Cosmos.Proto.SigningInput.create({
     publicKey: new Uint8Array(pubKeyData),
     signingMode: TW.Cosmos.Proto.SigningMode.Protobuf,
-    chainId: walletCore.CoinTypeExt.chainId(coinType),
+    chainId: 'mayachain-mainnet-v1',
     accountNumber: new Long(Number(chainSpecific.accountNumber)),
     sequence: new Long(Number(chainSpecific.sequence)),
     mode: TW.Cosmos.Proto.BroadcastMode.SYNC,
@@ -113,5 +97,5 @@ export const getThorPreSignedInputData: PreSignedInputDataResolver<
     }),
   })
 
-  return TW.Cosmos.Proto.SigningInput.encode(input).finish()
+  return [TW.Cosmos.Proto.SigningInput.encode(input).finish()]
 }
