+++ core/mpc/keysign/txInputData/thor/index.ts
@@ -1,46 +1,66 @@
+import { cosmosFeeCoinDenom } from '@core/chain/chains/cosmos/cosmosFeeCoinDenom'
 import { cosmosGasLimitRecord } from '@core/chain/chains/cosmos/cosmosGasLimitRecord'
 import { getCoinType } from '@core/chain/coin/coinType'
+import { isFeeCoin } from '@core/chain/coin/utils/isFeeCoin'
 import { assertField } from '@lib/utils/record/assertField'
 import { TW } from '@trustwallet/wallet-core'
 import Long from 'long'
 
-import { PreSignedInputDataResolver } from './PreSignedInputDataResolver'
+import { fromCommCoin } from '../../../types/utils/commCoin'
+import { TxInputDataResolver } from '../TxInputDataResolver'
 
-export const getMayaPreSignedInputData: PreSignedInputDataResolver<
-  'mayaSpecific'
-> = ({ keysignPayload, walletCore, chain, chainSpecific }) => {
+export const getThorTxInputData: TxInputDataResolver<
+  'thorchainSpecific'
+> = async ({ keysignPayload, walletCore, chain, chainSpecific }) => {
   const coinType = getCoinType({
     walletCore,
     chain,
   })
 
-  const coin = assertField(keysignPayload, 'coin')
+  const commCoin = assertField(keysignPayload, 'coin')
 
-  const fromAddr = walletCore.AnyAddress.createBech32(
-    coin.address,
-    coinType,
-    'maya'
+  const fromAddr = walletCore.AnyAddress.createWithString(
+    commCoin.address,
+    coinType
   )
 
-  const pubKeyData = Buffer.from(coin.hexPublicKey, 'hex')
+  const pubKeyData = Buffer.from(commCoin.hexPublicKey, 'hex')
 
   let thorchainCoin = TW.Cosmos.Proto.THORChainCoin.create({})
   let message: TW.Cosmos.Proto.Message[]
 
-  if (chainSpecific.isDeposit) {
+  if (keysignPayload.memo?.startsWith('merge:')) {
+    const fullDenom = keysignPayload.memo!.toLowerCase().replace('merge:', '')
+
+    message = [
+      TW.Cosmos.Proto.Message.create({
+        wasmExecuteContractGeneric:
+          TW.Cosmos.Proto.Message.WasmExecuteContractGeneric.create({
+            senderAddress: commCoin.address,
+            contractAddress: keysignPayload.toAddress,
+            executeMsg: '{ "deposit": {} }',
+            coins: [
+              TW.Cosmos.Proto.Amount.create({
+                denom: fullDenom,
+                amount: keysignPayload.toAmount,
+              }),
+            ],
+          }),
+      }),
+    ]
+  } else if (chainSpecific.isDeposit) {
     thorchainCoin = TW.Cosmos.Proto.THORChainCoin.create({
       asset: TW.Cosmos.Proto.THORChainAsset.create({
-        chain: 'MAYA',
-        symbol: 'CACAO',
-        ticker: 'CACAO',
+        chain: 'THOR',
+        symbol: commCoin.ticker,
+        ticker: commCoin.ticker,
         synth: false,
       }),
     })
-
     const toAmount = Number(keysignPayload.toAmount || '0')
     if (toAmount > 0) {
       thorchainCoin.amount = keysignPayload.toAmount
-      thorchainCoin.decimals = new Long(coin.decimals)
+      thorchainCoin.decimals = new Long(commCoin.decimals)
     }
 
     message = [
@@ -54,15 +74,14 @@ export const getMayaPreSignedInputData: PreSignedInputDataResolver<
       }),
     ]
   } else {
-    const toAddress = walletCore.AnyAddress.createBech32(
+    const toAddress = walletCore.AnyAddress.createWithString(
       keysignPayload.toAddress,
-      coinType,
-      'maya'
+      coinType
     )
-
-    if (toAddress.description() !== keysignPayload.toAddress) {
-      throw new Error('To address is different from the bech32 address')
+    if (!toAddress) {
+      throw new Error('invalid to address')
     }
+    const coin = fromCommCoin(commCoin)
 
     message = [
       TW.Cosmos.Proto.Message.create({
@@ -70,7 +89,7 @@ export const getMayaPreSignedInputData: PreSignedInputDataResolver<
           fromAddress: fromAddr.data(),
           amounts: [
             TW.Cosmos.Proto.Amount.create({
-              denom: coin.ticker.toLowerCase(),
+              denom: isFeeCoin(coin) ? cosmosFeeCoinDenom[chain] : coin.id,
               amount: keysignPayload.toAmount,
             }),
           ],
@@ -83,7 +102,7 @@ export const getMayaPreSignedInputData: PreSignedInputDataResolver<
   const input = TW.Cosmos.Proto.SigningInput.create({
     publicKey: new Uint8Array(pubKeyData),
     signingMode: TW.Cosmos.Proto.SigningMode.Protobuf,
-    chainId: 'mayachain-mainnet-v1',
+    chainId: walletCore.CoinTypeExt.chainId(coinType),
     accountNumber: new Long(Number(chainSpecific.accountNumber)),
     sequence: new Long(Number(chainSpecific.sequence)),
     mode: TW.Cosmos.Proto.BroadcastMode.SYNC,
@@ -94,5 +113,5 @@ export const getMayaPreSignedInputData: PreSignedInputDataResolver<
     }),
   })
 
-  return TW.Cosmos.Proto.SigningInput.encode(input).finish()
+  return [TW.Cosmos.Proto.SigningInput.encode(input).finish()]
 }
