+++ core/ui/mpc/keysign/action/mutations/useKeysignMutation.ts
@@ -10,11 +10,11 @@ import { getPreSigningHashes } from '@core/chain/tx/preSigningHashes'
 import { generateSignature } from '@core/chain/tx/signature/generateSignature'
 import { hexEncode } from '@core/chain/utils/walletCore/hexEncode'
 import { KeysignMessagePayload } from '@core/mpc/keysign/keysignPayload/KeysignMessagePayload'
+import { getTxInputData } from '@core/mpc/keysign/txInputData'
+import { getKeysignChain } from '@core/mpc/keysign/utils/getKeysignChain'
 import { useAssertWalletCore } from '@core/ui/chain/providers/WalletCoreProvider'
 import { useKeysignAction } from '@core/ui/mpc/keysign/action/state/keysignAction'
 import { customMessageConfig } from '@core/ui/mpc/keysign/customMessage/config'
-import { getTxInputData } from '@core/ui/mpc/keysign/tx/inputData/getTxInputData'
-import { getKeysignChain } from '@core/ui/mpc/keysign/utils/getKeysignChain'
 import { useCurrentVault } from '@core/ui/vault/state/currentVault'
 import { matchRecordUnion } from '@lib/utils/matchRecordUnion'
 import { chainPromises } from '@lib/utils/promise/chainPromises'
